Adds CfgConfigSynchronizer
 This service attempts to ensure that configurations created using other
means than file or config:edit, are replicated as a cfg file in
${karaf.etc} directory. This is mainly the case for configurations
created via web console.

The service listens for configuration changes and when a configuration
lacking the felix.fileinstall.filename property is detected, it will
create a cfg file with the configuration properties and then add a
felix.fileinstall.filename property.

Change-Id: I9c655c15e44383dc4847aff625e2696c7f216537